Integrate cmake & reorganize files
